+++ Engine/Include/State/State.h
@@ -16,7 +16,7 @@ namespace state
 		virtual ~State() = default;
 		virtual void Enter(T& owner) = 0;
 		virtual void Exit(T& owner) = 0;
-		virtual void Update(T& owner, float delta) = 0;
+		virtual void Update(T& owner, double delta) = 0;
 		virtual bool IsFinished(T& owner) = 0;
 	};
 }
