+++ Engine/Include/Timer/FrameRateController.h
@@ -6,23 +6,23 @@ namespace timer
 	class FrameRateController
 	{
 	private:
-		float m_targetInterval;
-		float m_elapsedTimeAccumulator;
+		double m_targetInterval;
+		double m_elapsedTimeAccumulator;
 
 		// TODO: remove this once IsReady() is removed
-		float m_lastElapsedTime;
+		double m_lastElapsedTime;
 
-		event::Event<float> FrameCompletedEvent;
+		event::Event<double> FrameCompletedEvent;
 
 	public:
-		explicit FrameRateController(float targetFPS = 60.0f) :
+		explicit FrameRateController(double targetFPS = 60.0f) :
 			m_targetInterval(1.0f / targetFPS),
 			m_elapsedTimeAccumulator(0.0f),
 			m_lastElapsedTime(0.0f)
 		{
 		}
 
-		void Update(float elapsedTime)
+		void Update(double elapsedTime)
 		{
 			m_elapsedTimeAccumulator += elapsedTime;
 
@@ -35,39 +35,39 @@ namespace timer
 		}
 
 		// subscribe as free function
-		void operator+=(event::Handler<void, void, float> handler)
+		void operator+=(event::Handler<void, void, double> handler)
 		{
 			FrameCompletedEvent += handler;
 		}
 
 		// unsubscribe as free function
-		void operator-=(event::Handler<void, void, float> handler)
+		void operator-=(event::Handler<void, void, double> handler)
 		{
 			FrameCompletedEvent -= handler;
 		}
 
 		// subscribe as class method
 		template <typename C>
-		void operator+=(event::Handler<void, C, float> handler)
+		void operator+=(event::Handler<void, C, double> handler)
 		{
 			FrameCompletedEvent += handler;
 		}
 
 		// unsubscribe as class method
 		template <typename C>
-		void operator-=(event::Handler<void, C, float> handler)
+		void operator-=(event::Handler<void, C, double> handler)
 		{
 			FrameCompletedEvent -= handler;
 		}
 
 		// subscribe as class lambda or std::function. TODO: test if this works
-		void operator+=(event::Handler<void, std::function<void(float)>, float> handler)
+		void operator+=(event::Handler<void, std::function<void(double)>, double> handler)
 		{
 			FrameCompletedEvent += handler;
 		}
 
 		// unsubscribe as class lambda or std::function. TODO: test if this works
-		void operator-=(event::Handler<void, std::function<void(float)>, float> handler)
+		void operator-=(event::Handler<void, std::function<void(double)>, double> handler)
 		{
 			FrameCompletedEvent -= handler;
 		}
@@ -77,7 +77,7 @@ namespace timer
 		// this method is a way to use this as polling, where listener polls if ready to render
 		// but i added a "subscription" method to align other timer class designs. 
 		// i should be removing this now as it is redundant, but let us keep for now...	
-		bool IsReady(float delta)
+		bool IsReady(double delta)
 		{
 			m_elapsedTimeAccumulator += delta;
 
@@ -93,7 +93,7 @@ namespace timer
 
 		// TODO:
 		// remove this to if removing IsReady. this only exist for IsReady
-		float GetLastElapsedTime() const
+		double GetLastElapsedTime() const
 		{
 			return m_lastElapsedTime;
 		}
